Separated the notebook permissions
Notebooks shared the save permissions with ML studio.
Separated them by introducing more granular permissions
Added migration script
