+++ backend/dataall/cdkproxy/stacks/policies/__init__.py
@@ -5,9 +5,5 @@ from dataall.cdkproxy.stacks.policies import (
     lakeformation, quicksight, redshift, stepfunctions, data_policy, service_policy
 )
 
-
-# TODO : should go away after refactoring of ml studios
-import dataall.modules.common.sagemaker.policies as sagemaker
-
 __all__ = ["_lambda", "cloudformation", "codestar", "databrew", "glue", "lakeformation", "quicksight",
-           "redshift", "stepfunctions", "data_policy", "service_policy", "sagemaker"]
+           "redshift", "stepfunctions", "data_policy", "service_policy", "mlstudio"]
