+++ backend/dataall/db/permissions.py
@@ -32,6 +32,7 @@ MANAGE_WORKSHEETS = 'MANAGE_WORKSHEETS'
 MANAGE_GLOSSARIES = 'MANAGE_GLOSSARIES'
 MANAGE_ENVIRONMENTS = 'MANAGE_ENVIRONMENTS'
 MANAGE_ORGANIZATIONS = 'MANAGE_ORGANIZATIONS'
+MANAGE_SGMSTUDIO_NOTEBOOKS = 'MANAGE_SGMSTUDIO_NOTEBOOKS'
 
 """
 ENVIRONMENT
@@ -282,6 +283,7 @@ TENANT_ALL = [
     MANAGE_GROUPS,
     MANAGE_ENVIRONMENTS,
     MANAGE_ORGANIZATIONS,
+    MANAGE_SGMSTUDIO_NOTEBOOKS,
 ]
 
 TENANT_ALL_WITH_DESC = {k: k for k in TENANT_ALL}
@@ -294,6 +296,7 @@ TENANT_ALL_WITH_DESC[MANAGE_ENVIRONMENTS] = 'Manage environments'
 TENANT_ALL_WITH_DESC[MANAGE_GROUPS] = 'Manage teams'
 TENANT_ALL_WITH_DESC[MANAGE_PIPELINES] = 'Manage pipelines'
 TENANT_ALL_WITH_DESC[MANAGE_ORGANIZATIONS] = 'Manage organizations'
+TENANT_ALL_WITH_DESC[MANAGE_SGMSTUDIO_NOTEBOOKS] = 'Manage ML studio notebooks'
 
 """
 REDSHIFT CLUSTER
