+++ backend/dataall/modules/notebooks/services/permissions.py
@@ -9,11 +9,15 @@ from dataall.db.permissions import (
     RESOURCES_ALL_WITH_DESC,
     RESOURCES_ALL,
     ENVIRONMENT_INVITATION_REQUEST,
+    TENANT_ALL,
+    TENANT_ALL_WITH_DESC
 )
 
 GET_NOTEBOOK = "GET_NOTEBOOK"
 UPDATE_NOTEBOOK = "UPDATE_NOTEBOOK"
 DELETE_NOTEBOOK = "DELETE_NOTEBOOK"
+CREATE_NOTEBOOK = "CREATE_NOTEBOOK"
+MANAGE_NOTEBOOKS = "MANAGE_NOTEBOOKS"
 
 NOTEBOOK_ALL = [
     GET_NOTEBOOK,
@@ -24,12 +28,21 @@ NOTEBOOK_ALL = [
 LIST_ENVIRONMENT_NOTEBOOKS = 'LIST_ENVIRONMENT_NOTEBOOKS'
 
 ENVIRONMENT_ALL.append(LIST_ENVIRONMENT_NOTEBOOKS)
+ENVIRONMENT_ALL.append(CREATE_NOTEBOOK)
 ENVIRONMENT_INVITED.append(LIST_ENVIRONMENT_NOTEBOOKS)
+ENVIRONMENT_INVITED.append(CREATE_NOTEBOOK)
 ENVIRONMENT_INVITATION_REQUEST.append(LIST_ENVIRONMENT_NOTEBOOKS)
+ENVIRONMENT_INVITATION_REQUEST.append(CREATE_NOTEBOOK)
 
+TENANT_ALL.append(MANAGE_NOTEBOOKS)
+TENANT_ALL_WITH_DESC[MANAGE_NOTEBOOKS] = "Manage notebooks"
+
+
+RESOURCES_ALL.append(CREATE_NOTEBOOK)
 RESOURCES_ALL.extend(NOTEBOOK_ALL)
 RESOURCES_ALL.append(LIST_ENVIRONMENT_NOTEBOOKS)
 
+RESOURCES_ALL_WITH_DESC[CREATE_NOTEBOOK] = "Create notebooks on this environment"
 RESOURCES_ALL_WITH_DESC[LIST_ENVIRONMENT_NOTEBOOKS] = "List notebooks on this environment"
 RESOURCES_ALL_WITH_DESC[GET_NOTEBOOK] = "General permission to list notebooks"
 RESOURCES_ALL_WITH_DESC[DELETE_NOTEBOOK] = "Permission to delete notebooks"
